Reorganise routes.handlers
