--- conman/routes/handlers/__init__.py
@@ -1,5 +1,0 @@
-from .base import BaseHandler
-from .simple import SimpleHandler
-
-
-__all__ = (BaseHandler, SimpleHandler)
