+++ conman/routes/simple/views.py
