+++ conman/routes/tests/test_handlers.py
@@ -11,7 +11,7 @@ class BaseHandlerPathTest(TestCase):
     """Test BaseHandler.path()."""
     def test_path(self):
         """Test directly on base class."""
-        base_handler_path = 'conman.routes.handlers.base.BaseHandler'
+        base_handler_path = 'conman.routes.handlers.BaseHandler'
         self.assertEqual(BaseHandler.path(), base_handler_path)
 
     def test_path_on_subclass(self):
